refactor: nextButton Click hooks 분리
